@@ -36,7 +36,7 @@
     "@types/express": "^4.17.2",
     "@types/request": "^2.48.2",
     "@types/ws": "^6.0.4",
-    "tslint": "^5.11.0",
-    "typescript": "^3.6.4"
+    "tslint": "~6.1.0",
+    "typescript": "~3.9.7"
   }
 }
