@@ -12,9 +12,10 @@ import { TimeSinceComponent } from '../components/time-since/time-since.componen
 import { ClipboardComponent } from '../components/clipboard/clipboard.component';
 import { QrcodeComponent } from '../components/qrcode/qrcode.component';
 import { FiatComponent } from '../fiat/fiat.component';
-import { NgbNavModule, NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
+import { NgbNavModule, NgbTooltipModule, NgbButtonsModule, NgbPaginationModule, NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
 import { TxFeaturesComponent } from '../components/tx-features/tx-features.component';
 import { TxFeeRatingComponent } from '../components/tx-fee-rating/tx-fee-rating.component';
+import { ReactiveFormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [
@@ -35,8 +36,12 @@ import { TxFeeRatingComponent } from '../components/tx-fee-rating/tx-fee-rating.
   ],
   imports: [
     CommonModule,
+    ReactiveFormsModule,
     NgbNavModule,
     NgbTooltipModule,
+    NgbButtonsModule,
+    NgbPaginationModule,
+    NgbDropdownModule,
   ],
   providers: [
     VbytesPipe,
@@ -44,7 +49,11 @@ import { TxFeeRatingComponent } from '../components/tx-fee-rating/tx-fee-rating.
   exports: [
     NgbNavModule,
     CommonModule,
+    ReactiveFormsModule,
     NgbTooltipModule,
+    NgbButtonsModule,
+    NgbPaginationModule,
+    NgbDropdownModule,
     TimeSinceComponent,
     ClipboardComponent,
     QrcodeComponent,
